Refactor offline code to a subclass
